+++ src/tk/PImageLoaderJpeg.hh
@@ -24,7 +24,7 @@
 namespace PImageLoaderJpeg
 {
 	const char *getExt(void);
-	uchar* load(const std::string &file, size_t &width, size_t &height,
+	uchar* load(const std::string &file, uint &width, uint &height,
 		    bool &use_alpha);
 }
