+++ src/tk/PImageLoaderXpm.hh
@@ -27,7 +27,7 @@ extern "C" {
 namespace PImageLoaderXpm
 {
 	const char *getExt(void);
-	uchar* load(const std::string &file, size_t &width, size_t &height,
+	uchar* load(const std::string &file, uint &width, uint &height,
 		    bool &use_alpha);
 }
